commit29
Remove anonymous self-invoked function from game.js
